backend/lending-service/pyproject.toml
[tool.poetry]
name = "lending-service"
version = "0.1.0"
description = ""
authors = ["Nikodem <nikodemmlynski08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^6.0.3"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.2"
requests = "^2.32.5"
pyjwt = "^2.11.0"
cryptography = "^46.0.5"
celery = "^5.4.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
